restructure list calls
